+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added conversion between Katakana and Cyrillic scripts.
+- Added options for ィ (-y), ゥ (-w) and ㇴ (-n) usage in Latin to Katakana conversion.
 - Added support from Python 3.8 and above.
 - Added more badges to [README.md](README.md).
 - Added [tox](https://tox.readthedocs.io/) for multi-environment testing.